Fix type casting
