+++ src/modules/Elsa.CSharp/Handlers/ConfigureWorkflowVariables.cs
@@ -14,7 +14,7 @@ namespace Elsa.CSharp.Handlers;
 public class ConfigureWorkflowVariables : INotificationHandler<EvaluatingCSharp>
 {
     /// <inheritdoc />
-    public async Task HandleAsync(EvaluatingCSharp notification, CancellationToken cancellationToken)
+    public Task HandleAsync(EvaluatingCSharp notification, CancellationToken cancellationToken)
     {
         var expressionExecutionContext = notification.Context;
         var variables = expressionExecutionContext.GetVariablesInScope().ToList();
@@ -27,9 +27,10 @@ public class ConfigureWorkflowVariables : INotificationHandler<EvaluatingCSharp>
         {
             var variableName = variable.Name.Pascalize();
             var variableType = variable.GetVariableType();
-            sb.AppendLine($"\tpublic {variableType.FullName} {variableName}");
+            var friendlyTypeName = GetFriendlyTypeName(variableType);
+            sb.AppendLine($"\tpublic {friendlyTypeName} {variableName}");
             sb.AppendLine("\t{");
-            sb.AppendLine($"\t\tget => Context.GetVariable<{variableType.FullName}>(\"{variableName}\");");
+            sb.AppendLine($"\t\tget => Context.GetVariable<{friendlyTypeName}>(\"{variableName}\");");
             sb.AppendLine($"\t\tset => Context.SetVariable(\"{variableName}\", value);");
             sb.AppendLine("\t}");
         }
@@ -37,5 +38,27 @@ public class ConfigureWorkflowVariables : INotificationHandler<EvaluatingCSharp>
         sb.AppendLine("}");
         sb.AppendLine("var Variables = new WorkflowVariablesWrapper(Context);");
         notification.AppendScript(sb.ToString());
+        return Task.CompletedTask;
+    }
+    
+    static string GetFriendlyTypeName(Type type)
+    {
+        if (!type.IsGenericType)
+            return type.FullName!;
+
+        var sb = new StringBuilder();
+        sb.Append(type.Namespace);
+        sb.Append('.');
+        sb.Append(type.Name.Substring(0, type.Name.IndexOf('`')));
+        sb.Append('<');
+        var genericArgs = type.GetGenericArguments();
+        for (var i = 0; i < genericArgs.Length; i++)
+        {
+            if (i > 0)
+                sb.Append(", ");
+            sb.Append(GetFriendlyTypeName(genericArgs[i]));
+        }
+        sb.Append('>');
+        return sb.ToString();
     }
 }
