+++ src/modules/Elsa.Workflows.Core/Models/ActivityInput.cs
@@ -1,0 +1,6 @@
+namespace Elsa.Workflows.Core.Models;
+
+/// <summary>
+/// Represents an activity input name and value.
+/// </summary>
+public record WorkflowInput(string Name, object? Value);
